+++ src/main/java/com/knockbook/backend/dto/ProductDetailDTO.java
@@ -4,8 +4,6 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -18,11 +16,9 @@ public class ProductDetailDTO {
     private String manufacturerName;
     private String isImported;
     private String importCountry;
-    private BigDecimal averageRating;
+    private Double averageRating;
     private Integer reviewCount;
 
-    // GALLERY: 최대 4장, sort_order 1..4
     private List<String> galleryImageUrls;
-    // DESCRIPTION: 본문 이미지 전부
     private List<String> descriptionImageUrls;
 }
