+++ src/main/java/com/knockbook/backend/dto/ProductSummaryDTO.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder
@@ -14,7 +12,7 @@ public class ProductSummaryDTO {
     private String name;
     private Integer unitPriceAmount;
     private Integer salePriceAmount;
-    private BigDecimal averageRating;
+    private Double averageRating;
     private Integer reviewCount;
     private String thumbnailUrl;
     private String availability;
